refactor: refactor several services using solid
This commit contains several updates, they occured by using SRP(Single
responsibility principle)
