+++ src/email/interfaces/mailing-service.interface.ts
@@ -2,4 +2,5 @@ import { ISendMail } from './send-mail.interface';
 
 export interface IMailingService {
   sendMail(payload: ISendMail): Promise<void>;
+  sendMailsToSubscribers(): Promise<void>;
 }
