--- src/subscription/interfaces/send-to-subscribers.interface.ts
@@ -1,5 +1,0 @@
-export interface ISendToSubscribers {
-  readonly html: string;
-
-  readonly subject: string;
-}
